hotfix(menu_module): fix bug in handler of module menu
