+++ main.c
@@ -1,6 +1,7 @@
 #include "raylib.h"
-// #include "src/modules/menu/menu.h"
-#include "menu.h"
+#include "./src/modules/menu/menu.h"
+
+// #include "menu.h"
 
 int main() {
     menuScreen();
