--- menu.h
@@ -1,6 +1,0 @@
-#ifndef MENU_H
-#define MENU_H
-
-void menuScreen();
-
-#endif
